+++ src/lib.rs
@@ -6,7 +6,6 @@
 use std::sync::OnceLock;
 
 pub mod bundle;
-pub mod bundle_index;
 pub mod bundle_loader;
 pub mod commands;
 pub mod dat;
